+++ database/migrations/2023_04_04_011749_create_laporan_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('laporan', function (Blueprint $table) {
             $table->id('Id_Laporan');
-            $table->integer('Id_Barang');
+            $table->unsignedBigInteger('Id_Barang');
             $table->integer('Harga_Beli');
             $table->integer('Harga_Jual');
             $table->integer('Kuantitas');
