+++ database/migrations/2023_05_25_005421_create_pesanan_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('pesanan', function (Blueprint $table) {
             $table->id('Id_Pesanan');
-            $table->integer('Id_Pelanggan');
+            $table->unsignedBigInteger('Id_Pelanggan');
             $table->integer('Total_Harga');
             $table->date('Tgl_Pembayaran');
             $table->enum('Metode_Pembayaran', ['Transfer Bank','COD']);
